# swagger.yml
openapi: 3.0.0
info:
  title: Quinnn API List
  description: This is a reference api list of Quinnn Backend.
  version: 1.0.0
  contact:
    email: ankuragarwa246@gmail.com
servers:
  - url: https://lgbtq-backend-free.azurewebsites.net/api
    description: Production Server (Azure)
  - url: http://localhost:9999/api
    description: Development server
paths:
...
  /post/create:
    post:
      tags:
        - Post
      summary: Create a post
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/CreatePost'
      responses:
        '200':
          description: Post Created Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /post/all:
    get:
      tags:
        - Post
      summary: Get all posts of the logged in user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Posts Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /post/update/{postId}:
    put:
      tags:
        - Post
      summary: Update a particular post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          description: Post ID
          schema:
            type: string
            format: ObjectId
      requestBody:
        $ref: '#/components/requestBodies/UpdatePost'
      responses:
        '200':
          description: Post Updated Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /post/delete/{postId}:
    delete:
      tags:
        - Post
      summary: Delete a particular post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          description: Post ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Post Deleted Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /post/like/{postId}:
    put:
      tags:
        - Post
      summary: Like a post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          description: Post ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Post Liked Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /post/unlike/{postId}:
    put:
      tags:
        - Post
      summary: Unlike a post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          description: Post ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Post Unliked Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /event/create:
    post:
      tags:
        - Event
      summary: Create an event
      requestBody:
        $ref: '#/components/requestBodies/CreateEvent'
      responses:
        '200':
          description: Event Created Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /event/fetch/{eventId}:
    get:
      tags:
        - Event
      summary: Get a particular event
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: eventId
          description: Event ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Event Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /event/update/{eventId}:
    put:
      tags:
        - Event
      summary: Update an event
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: eventId
          description: Event ID
          schema:
            type: string
            format: ObjectId
      requestBody:
        $ref: '#/components/requestBodies/UpdateEvent'
      responses:
        '200':
          description: Event Updated Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /event/delete/{eventId}:
    delete:
      tags:
        - Event
      summary: Delete an event
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: eventId
          description: Event ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Event Deleted Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /event/register/{eventId}:
    put:
      tags:
        - Event
      summary: Register for an event
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: eventId
          description: Event ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Event Registered Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
components:
  schemas:
...
    CreatePostInput:
      type: object
      properties:
        text:
          type: string
        imageURL:
          type: string
    UpdatePostInput:
      type: object
      properties:
        text:
          type: string
...
    CreateEventInput:
      type: object
      properties:
        title:
          type: string
          required: true
        description:
          type: string
          required: true
        isOnline:
          type: boolean
          default: true
        location:
          type: string
        meetingLink:
          type: string
        dateTime:
          type: string
          format: date-time
          required: true
        imageURL:
          type: string
    UpdateEventInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        isOnline:
          type: boolean
        location:
          type: string
        meetingLink:
          type: string
        dateTime:
          type: string
          format: date-time
        imageURL:
          type: string
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
...
    CreatePost:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreatePostInput'
        description: Create Post
    UpdatePost:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdatePostInput'
        description: Update Post
...
    CreateEvent:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateEventInput'
        description: Create Event
    UpdateEvent:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateEventInput'
        description: Update Event